DAY_1/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
faiss-cpu
sentence-transformers
pypdf
python-dotenv
chromadb
langchain-google-genai
